refactor: move generic url reading to Utils
