+++ src/main/java/dinkplugin/util/Utils.java
@@ -1,13 +1,20 @@
 package dinkplugin.util;
 
+import com.google.gson.Gson;
+import com.google.gson.reflect.TypeToken;
 import lombok.experimental.UtilityClass;
 import net.runelite.api.Client;
 import net.runelite.api.vars.AccountType;
 import net.runelite.api.widgets.Widget;
 import net.runelite.api.widgets.WidgetInfo;
 import net.runelite.client.util.ColorUtil;
+import okhttp3.Call;
+import okhttp3.Callback;
 import okhttp3.MediaType;
 import okhttp3.MultipartBody;
+import okhttp3.OkHttpClient;
+import okhttp3.Request;
+import okhttp3.Response;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
@@ -16,6 +23,9 @@ import java.awt.Color;
 import java.awt.image.BufferedImage;
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
+import java.io.Reader;
+import java.util.concurrent.CompletableFuture;
+import java.util.function.Function;
 
 @UtilityClass
 public class Utils {
@@ -63,6 +73,34 @@ public class Utils {
             MediaType type = part.body().contentType();
             return type != null && "image".equals(type.type());
         });
+    }
+
+    public <T> CompletableFuture<T> readJson(@NotNull OkHttpClient httpClient, @NotNull Gson gson, @NotNull String url, @NotNull TypeToken<T> type) {
+        return readUrl(httpClient, url, reader -> gson.fromJson(reader, type.getType()));
+    }
+
+    public <T> CompletableFuture<T> readUrl(@NotNull OkHttpClient httpClient, @NotNull String url, @NotNull Function<Reader, T> transformer) {
+        CompletableFuture<T> future = new CompletableFuture<>();
+        Request request = new Request.Builder().url(url).build();
+        httpClient.newCall(request).enqueue(new Callback() {
+            @Override
+            public void onFailure(@NotNull Call call, @NotNull IOException e) {
+                future.completeExceptionally(e);
+            }
+
+            @Override
+            public void onResponse(@NotNull Call call, @NotNull Response response) {
+                assert response.body() != null;
+                try (Reader reader = response.body().charStream()) {
+                    future.complete(transformer.apply(reader));
+                } catch (Exception e) {
+                    future.completeExceptionally(e);
+                } finally {
+                    response.close();
+                }
+            }
+        });
+        return future;
     }
 
 }
